bots/bot_numbers/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
APScheduler==3.11.0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
gspread-formatting==1.2.1
httplib2==0.22.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
